v2v: -o rhv-upload: Don't require ‘-of raw’ parameter.
Because we checked the limitation of raw+sparse during command line
processing, it had the effect of forcing you to use ‘-of raw’ even if
the input was already in raw format.  Move the checking to the output
mode to avoid this.

Fixes commit cc04573927cca97de60d544d37467e67c25867a7.