v2v: Use -ip to pass input password (instead of --password-file).
Consistent with the option -op added in
commit a4e181137a38f5767dd1bf05dc482959cb7283be, use -ip to pass
passwords for the input side.

This replaces --password-file, and like that option you have to pass a
filename.

This also changes the code so we pass around the filename instead of
the password between functions.  For ‘-it vddk’ this simplifies
things: we can pass the filename through to nbdkit and we never need
to read it in virt-v2v.  For other input methods we eventually need to
read the password from the file at some lower level place in the code.