v2v: Add output#prepare_metadata method to prepare metadata before copying.
Currently unused, this method will allow output methods to prepare the
target hypervisor to receive a guest before the actual copy/upload
begins.

For rhv-upload this is useful since RHV would prefer to have the
metadata for the guest before receiving disk data.  The only metadata
unavailable here is target_stats.target_actual_size, but in the RHV API
case this is never available as we stream the data directly into RHV
rather than writing it to a local file.