v2v: Move output#prepare_targets call to just before the copy.
Previously we initialized the targets and called #prepare_targets very
early on.  However this causes various problems:

 - the targets struct is only half-initialized
   (target_file is set to a dummy value)
 - the targets struct is passed around unnecessarily

The ostensible reason for this was so that #prepare_targets is called
very early on so that errors can be displayed quickly.  This is
important, but in practice it doesn't really matter if the user has to
wait for a few seconds for the conversion to happen first.

We also have the #precheck method which can do some pre-checking
functions, and that does run very early.

This is just refactoring, but it will allow some more significant
changes in how targets get initialized.  There are some large
whitespace changes in this patch so it is better to look at it using
‘git show -w’.