p2v: turn Reboot button into a Shutdown popup menu button
To ease adding a proper Shutdown action in the conversion dialog, turn
the current Reboot button into a button that pops a menu up, with Reboot
as the only available action.

The menu is implemented using a popover when using "recent" versions of
GLib, and GTK+ 3, because of all the APIs needed.  In older versions, a
simple toggle button with menu is used, which though needs a manual menu
firing when pressing the button itself (and not its side arrow).