v2v: -o rhv-upload: Fix test so it doesn't fail if Content-Length header missing.
If the Content-Length header was missing from the headers returned by
the server then the test would fail with:

KeyError: 'content-length'

This happens on RHEL 7.6 in particular.  We can fix this by using an
alternate method to fetch the header, which will return None instead
of throwing an exception if the header doesn't exist.  See:
https://stackoverflow.com/a/19255675