v2v: docs: Remove sentence about supporting qemu-kvm-rhev in RHEL (RHBZ#1651977).
The previous support documentation said that for UEFI guests, when
using RHEL ≥ 7.3 host, you must use qemu-kvm-rhev.  For RHEL 8 this is
inaccurate since qemu-kvm-rhev no longer exists.

To fix this I've dropped the whole sentence.  It only applies to the
downstream product (RHEL) and so RHEL can add the right documentation
if it needs to.

Thanks: Ming Xie, Pino Toscano.