v2v: Add support for libosinfo metadata
There's a standardized libosinfo namespace for libvirt domain metadata.  For now
it supports the id of the OS only.  However that is still a very helpful feature
that is already supported in gnome-boxes and virt-manager (at least).

The discussion happened here:

  https://www.redhat.com/archives/libosinfo/2018-September/msg00003.html

So let's add the support to local and libvirt outputs.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>