rhv-upload: Fix cleanup after errors
When request failed, we paused the transfer. This is not needed since
our intent it to cancel the transfer.

When closing after failure, we canceled the transfer and removed the
disk. This is not needed since the transfer owns the disk and will
remove it when canceled.

When finalizing times out, we canceled the transfer and removed the
disk. This is not needed since the transfer will clean it self, and
likely to fail because cancelling is not possible after finalizing.