rhv-upload: Get host before creating disk
If getting the host object raised, we forgot to remove the disk. Getting
the host object before creating the disk avoids the error handling, and
will fail faster on errors, without waiting until the disk is ready.

This will help to extract the code for creating transfer out of open().