rhv-upload: Extract finalize_transfer() function
Extract finalize_transfer() function hiding the complexity of finalizing
and waiting for a transfer, and separating plugin logic like writing the
disk_id file from oVirt logic.

Since canceling a transfer is not part of finalizing, but the current
plugin policy, move cancellation after finalize failure to close().

When waiting for the disk status change, handle the ILLEGAL status, so
we can fail quickly if oVirt failed to finalized and paused the
transfer, instead of waiting for timeout.