rhv-upload: Support qcow2 disk format
When using oVirt >= 4.3, we can enable the NBD based backend in imageio
by specifying that we transfer raw data when creating a transfer. With  
the NBD backend, we can import to disks using qcow2 format.

To make it work, we override output#transfer_format to return always raw
format, but we create the disk on RHV side using qcow2 format.

The pipeline looks like this:

    qemu-img convert <nbd> rhv-upload <http> imageio <nbd> qemu-nbd