Revamp check for Python code style
It turns out that
a) the standalone pep8 executable is deprecated in favour of pycodestyle
b) $python-pep8 does not exist in modern distros
and thus the style checks for Python scripts have been skipped for a
long time.

Instead, switch to pycodestyle directly, updating the list of ignored
issues to the two categories left.