v2v: try to get windows driver files from libosinfo
Query libosinfo to get the list of files for drivers of Windows guests,
copying them if they are local files only. In case it is not possible
(e.g. no driver files are listed, or their location is a non-local URL,
or they do not exists), the manual filtering of the virtio-win content
is done as before.