v2v: Add standalone nbdkit module.
This standalone library can be used to examine nbdkit on the system,
probe for plugins and filters, and construct and run nbdkit commands.

We use this to rewrite and simplify both the Nbdkit_sources module and
the -o rhv-upload mode.

This is just refactoring, introducing common code for running nbdkit
in all the places where it is currently used.