tests: fix location to generated images
Some of the libvirt XMLs for tests refer to a generated phony disk
image, and so far the relative path used worked almost by chance.
(For each disk, the path to it was ../test-data/etc, and since the
overlay pointing to it is stored directly in $LIBGUESTFS_CACHEDIR,
which is $buildir/tmp, then the relative path was resolved.)

Instead, have configure place the right top-level directory in those
XMLs, so the full path points to the generated disks. This fixes the
tests for at least two cases:
- we change the place where the temporary overlays are stored
- in srcdir!=builddir builds

There is no functional change to the tests themselves.