build: Use libtool instead of $(LIBTOOL) in VG macro.
In commit 5f355c2952729581f1b988297fe8cabe756c01e2 I incorrectly
rewrote ‘libtool’ from the original libguestfs file as ‘$(LIBTOOL)’.

This was because of a comment that appears in the libguestfs file:

  Note we run libtool, not $(LIBTOOL) since the latter expands to
  libtool-kill-dependency-libs.sh

Since virt-v2v doesn't have libtool-kill-dependency-libs.sh, I
reasoned that we should change libtool back to $(LIBTOOL) and remove
the comment.

However the result of this was wrong.  $(LIBTOOL) expands to
$(top_builddir)/libtool, but this breaks (because of the relative
path) whenever the VG macro is used within pushd...popd pairs, as
happens in several tests.

Restoring the libguestfs version fixes this.