v2v: -o libvirt: Remove cache=none (RHBZ#1837453).
Traditionally if you did live migration (KVM to KVM), you had to
ensure that cache=none was set on all disks of the guest up front.
This was because of quirks in how NFS works (I think the close-to-open
consistency and the fact that during live migration both qemus have
the file open), and we had to assume the worst case that a guest might
be backed by NFS.

Because of this when virt-v2v converts a guest to run on KVM using
libvirt it sets cache=none.

This is not necessary with modern qemu.  If qemu supports the
drop-cache property of the file block driver, which libvirt will
automatically detect for us, then libvirt live migration is able to
tell qemu to drop cached data at the right time even if the backing is
NFS.

It also had a significant performance impact.  In some synthetic
benchmarks it could show 2 or 3 times slower performance.

Thanks: Ming Xie, Peter Krempa.
(cherry picked from commit 9720f45e0cd9283739fd2a67c19e66912489dfc7)