v2v: -it vddk: Don't use nbdkit readahead filter with VDDK (RHBZ#1832805).
This filter deliberately tries to coalesce reads into larger requests.
Unfortunately VMware has low limits on the size of requests it can
serve to a VDDK client and the larger requests would break with errors
like this:

  nbdkit: vddk[3]: error: [NFC ERROR] NfcFssrvrProcessErrorMsg: received NFC error 5 from server: Failed to allocate the requested 33554456 bytes

We already increase the maximum request size by changing the
configuration on the VMware server, but it's not sufficient for VDDK
with the readahead filter.

As readahead is only an optimization, the simplest solution is to
disable this filter when we're using nbdkit-vddk-plugin.

Thanks: Ming Xie