v2v: Remove extraneous '=' when setting bandwidth (RHBZ#1841096).
Commit c3a54d6aed6dfc65f9ffa59976bb8d20044c03a8 ("v2v: Add standalone
nbdkit module.") was supposed to be a simple refactoring but it broke
the --bandwidth and --bandwidth-file options (amongst other things).

Because of an extra '=' character which was accidentally left over, it
would add an extra character in the nbdkit-rate-filter command line.
For example:

  virt-v2v .. --bandwidth 200M

would invoke:

  nbdkit .. --filter rate rate==200M

which causes a parse error.  The --bandwidth-file option does not
invoke a parse error but does not work, for similar reasons.

Thanks: Ming Xie
(cherry picked from commit a89a084b2d0f6d40716c1d34969f6c49ea28e9b3)