v2v: Allow output to block devices (RHBZ#1868690).
We previously implicitly supported writing to block devices instead of
local files, but there were several problems:

* Block devices could be deleted, especially if virt-v2v failed during
  a conversion.

* Block devices could be overwritten by a file with the same name,
  although I believe this is just an observed consequence of the
  previous point, or at least I was not able to reproduce this until
  virt-v2v failed for another reason and then I noticed that because
  the block device was deleted, the next run overwrote it with a file.

* It was not documented anywhere how to do it.

This commit makes the small code change needed to allow virt-v2v to
write to a block device, only for existing outputs which write to
local files (ie. using TargetFile).  Also it avoids deleting block
devices accidentally on failure.

Note this commit intentionally does not prevent you from writing qcow2
to a block device.  RHV uses this so it is a thing that people do.