ovf: Try to set BiosType correctly according to the oVirt sources.
Previously we mapped TargetBIOS to 0 and TargetUEFI to 2, but I
believe that both of these are wrong, and they certainly don't match
the oVirt source code.

This changes the BiosType calculation to use all necessary inputs
(architecture, I440FX|Q35, and TargetBIOS|TargetUEFI) and try to map
these as closely as possible to the BiosType as described in the oVirt
sources.

An exact mapping is not possible, so in cases where we can't find one
we punt on this by returning 0 (use cluster default).