v2v: rhv-upload-plugin: Remove unsafe highestwrite
We had a mechanism to ignore zero writes if the written area in the disk
was never written. This is safe when working sparse files which are
always zeroed. When working with block devices, we don't have any
guarantee on the contents of the device, so we should not skip zero
requests.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>