v2v: rhv-upload-plugin: Support multiple connections
Use multiple connections to imageio server to speed up the transfer.

Connections are managed via a thread safe queue. Threads remove
a connection from the queue for every request, and put it back when at
the end of the request. Only one thread can access the connection at the
same time.

Threads are accessing existing values in the handle dict, like
h["path"]. They may also modify h["failed"] on errors. These operations
are thread safe and do not require additional locking.

Sending flush request is more tricky; imageio server does not guarantee
consistency for multiple for connections, so we need to send flush
command for every connection.

Since flush may be called when there are in-flight requests, we wait
until all in-flight request are done before sending a flush. While
flushing, the connection pool is empty so new requests will block on the
queue until the flush is completed.

Closing is done in a similar way, waiting until all in-flight requests
are done and closing all connections.

Testing shows that we requests are spread over 4 connections, but
performance is worse. Connection time increased from 5.36 seconds to
7.08 seconds (32% slower).

[connection 1 ops, 7.053520 s]
[dispatch 548 ops, 1.188910 s]
[write 469 ops, 1.014646 s, 332.75 MiB, 327.95 MiB/s]
[zero 77 ops, 0.058495 s, 1.13 GiB, 19.29 GiB/s]
[flush 2 ops, 0.000291 s]

[connection 1 ops, 7.085039 s]
[dispatch 548 ops, 1.097437 s]
[write 478 ops, 0.924214 s, 323.25 MiB, 349.76 MiB/s]
[zero 68 ops, 0.052265 s, 1.22 GiB, 23.43 GiB/s]
[flush 2 ops, 0.000258 s]

[connection 1 ops, 7.037253 s]
[dispatch 547 ops, 1.111386 s]
[write 477 ops, 0.959592 s, 343.25 MiB, 357.70 MiB/s]
[zero 68 ops, 0.047005 s, 1.20 GiB, 25.44 GiB/s]
[flush 2 ops, 0.000266 s]

[connection 1 ops, 7.045538 s]
[dispatch 548 ops, 1.191029 s]
[write 482 ops, 1.041125 s, 347.12 MiB, 333.41 MiB/s]
[zero 64 ops, 0.045171 s, 1.14 GiB, 25.16 GiB/s]
[flush 2 ops, 0.000210 s]

Signed-off-by: Nir Soffer <nsoffer@redhat.com>