v2v: rhv-upload-plugin: Support FUA
imageio does not have a FUA flag, but we can emulate it using the flush
feature. With flush=y, write or zero will issue a NBD_CMD_FLUSH at the
end, so the call returns only after the data is flushed to storage.

The plugin reports now that it supports FUA (based on can_flush), and
when FUA flag is set, it emulates fua by setting flush query (PUT) or
flag (PATCH).

For example, this nbd command:

    NBD_CMD_WRITE flags=NBD_CMD_FUA_FLAG

Is translated to this http request:

    PUT /path?flush=y

On imageio server, this translates back to:

    NBD_CMD_WRITE flags=0
    NBD_CMD_FLUSH

Imageio server uses preallocated buffer per connection. If the write
request is bigger than the buffer, it sends multiple write commands. The
request returns only when the flush is completed.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>