v2v: support configuration of viosock driver
Check that install_drivers function has copied viosock driver files to
the windows guest file system. If positive, this means the drivers have
passed minor/major version check and the guest is able to use them.
After we know that that the drivers are on the guest, we can enable
virtio sock option in configuration and starting script files.

Signed-off-by: Valeriy Vdovin <valeriy.vdovin@virtuozzo.com>