v2v: Increase required free space in Windows to 100 MB
With an increasing number of drivers being installed in Windows the
existing limit (20 MB) was far too low.  In fact we found that a guest
with 63 MB of free space would sometimes run out of space.

This commit increases the required space to 100 MB for Windows.

There are also a couple of smaller fixes:

 - We now properly distinguish between / as a Linux boot drive,
   and Windows.

 - The error message has been improved to display MBs instead of bytes.

Reported-by: Ming Xie
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1949147
(cherry picked from commit dfcf60c916a986a352938b432231a36558a3bc05)