v2v: -i ova: Set LANG=C and --no-auto-compress when running tar.
https://listman.redhat.com/archives/libguestfs/2021-April/msg00072.html

In non-English locales, the output from tar is translated (eg. in
French "block" -> "bloc").  This prevented us from parsing the output
of the tar command when looking for file offsets.

We should really be using nbdkit-tar-filter here, but in the meantime
I made sure that the environment and parameters to tar match.

Reported-by: Étienne BERSAC