rhv-upload: Validate UUIDs passed to -oo rhv-disk-uuid (RHBZ#1789279)
The validation helps us fail early and with a sensible error message.  The NIL
UUID is not valid for oVirt, but other than that there is no other logic in
there merely because the UUID types are a matter of the generator and they are
just forwarded in this partucular case.

On top of that also check that the UUID is not already taken.  This makes the
code fail with a sensible error message instead of cryptic error from ovirtsdk.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
(cherry picked from commit 5a0662b1d85ef4cd07a0da937730f4b08f539989)