v2v: Don't update source.s_name field
To implement the -on option we used essentially a hack of updating the
source.s_name field and using that field in output drivers etc.  This
saved us passing around a parameter at the cost of making the source
struct a kind of (not literally) mutable object.

Since we'd like to save source into a file in future and not have to
edit that file, reimplement this without mutating source.