v2v: Add s_disks field to source struct
This refactoring puts the disks where they belong in the same struct
with other source attributes.

This essentially reverts commit 0884a7f42a ("v2v: Split out source and
source_disks").  We will be splitting virt-v2v in a different way from
how we imagined it when we wrote that commit.