v2v: Remove overlays.ov_stats.target_estimated_size
This estimate wasn't very good.  It was only used for two purposes,
firstly to print an estimated size in debugging output, and secondly
to estimate the ovf:actual_size in OVF output in cases where we didn't
have an actual size.

In modular virt-v2v the OVF usage can be replaced by a true size in
all cases (since the output will always be an NBD endpoint so we can
examine it for extents etc).