tests: Split -oa preallocated test into qcow2 and raw
Inside qemu there are fairly different data paths for qcow2 and raw
files, so it makes sense to test -oa preallocated + -of qcow2|raw
separately.

The original test is just duplicated with small adjustments.  Both new
tests fail because of breakage in qemu:
https://www.mail-archive.com/qemu-devel@nongnu.org/msg671142.html
but I hope to fix this in modular virt-v2v, at least in the raw case.