v2v: -o rhv-upload: Generate disk UUIDs
With -o rhv-upload we can either use -oo rhv-disk-uuid to specify
UUIDs per disk, or without.  Previously not specifying caused us to
ask RHV to generate the UUIDs.  However this was awkward because we
had to communicate the UUIDs back to the main program through
temporary files.

Instead of doing that, if not specified, generate UUIDs using "uuidgen -r".

Idea suggested by Nir Soffer.